Adicionando fragment altera Criando AppModule Importando DataBindingConverters.kt em fragment_altera.xml e fragment_cadastra.xml Adicionando dependências do gradle Adicionando anotações do Hilt

+++ app/src/main/java/com/example/segundaprova/converters/DataBindingConverters.kt
@@ -15,7 +15,7 @@ class DataBindingConverters {
         }
 
         @JvmStatic
-        fun convertIntToString(value: Float?): String {
+        fun convertIntToString(value: Int?): String {
             return value?.toString() ?: ""
         }
     }
